Refactored IndexUtil and related test cases (#884)
