+++ src/main/java/io/anserini/search/SimpleSearcher.java
@@ -88,6 +88,7 @@ public class SimpleSearcher implements Closeable {
       new Sort(SortField.FIELD_SCORE,
           new SortField(TweetGenerator.StatusField.ID_LONG.name, SortField.Type.LONG, true));
   private static final Logger LOG = LogManager.getLogger(SimpleSearcher.class);
+
   private final IndexReader reader;
   private Similarity similarity;
   private Analyzer analyzer;
@@ -97,7 +98,7 @@ public class SimpleSearcher implements Closeable {
 
   private IndexSearcher searcher = null;
 
-  protected class Result {
+  public class Result {
     public String docid;
     public int ldocid;
     public float score;
@@ -184,21 +185,6 @@ public class SimpleSearcher implements Closeable {
     reader.close();
   }
 
-  public Result[] search(String q) throws IOException {
-    return search(q, 10);
-  }
-
-  public Result[] search(String q, int k) throws IOException {
-    return search(q, k, -1);
-  }
-
-  public Result[] search(String q, int k, long t) throws IOException {
-    Query query = new BagOfWordsQueryGenerator().buildQuery(LuceneDocumentGenerator.FIELD_BODY, analyzer, q);
-    List<String> queryTokens = AnalyzerUtils.tokenize(analyzer, q);
-
-    return search(query, queryTokens, q, k, t);
-  }
-
   public Map<String, Result[]> batchSearch(List<String> queries, List<String> qids, int k, int threads) {
     return batchSearch(queries, qids, k, -1, threads);
   }
@@ -223,7 +209,7 @@ public class SimpleSearcher implements Closeable {
         Long lineNumber = index.incrementAndGet();
         if (lineNumber % 100 == 0) {
           double timePerQuery = (double) (System.nanoTime() - startTime) / (lineNumber + 1) / 1e9;
-          System.out.format("Retrieving query " + lineNumber + " (%.3f s/query)\n", timePerQuery);
+          LOG.info(String.format("Retrieving query " + lineNumber + " (%.3f s/query)\n", timePerQuery));
         }
       });
     }
@@ -251,6 +237,21 @@ public class SimpleSearcher implements Closeable {
     return results;
   }
 
+  public Result[] search(String q) throws IOException {
+    return search(q, 10);
+  }
+
+  public Result[] search(String q, int k) throws IOException {
+    return search(q, k, -1);
+  }
+
+  public Result[] search(String q, int k, long t) throws IOException {
+    Query query = new BagOfWordsQueryGenerator().buildQuery(LuceneDocumentGenerator.FIELD_BODY, analyzer, q);
+    List<String> queryTokens = AnalyzerUtils.tokenize(analyzer, q);
+
+    return search(query, queryTokens, q, k, t);
+  }
+
   protected Result[] search(Query query, List<String> queryTokens, String queryString, int k, long t) throws IOException {
     // Initialize an index searcher only once
     if (searcher == null) {
@@ -263,7 +264,7 @@ public class SimpleSearcher implements Closeable {
     searchArgs.hits = k;
     searchArgs.searchtweets = searchtweets;
 
-    TopDocs rs = new TopDocs(new TotalHits(0, TotalHits.Relation.EQUAL_TO), new ScoreDoc[]{});
+    TopDocs rs;
     RerankerContext context;
     if (searchtweets) {
       if (t > 0) {
@@ -283,7 +284,7 @@ public class SimpleSearcher implements Closeable {
       }
     } else {
       rs = searcher.search(query, isRerank ? searchArgs.rerankcutoff : k, BREAK_SCORE_TIES_BY_DOCID, true);
-        context = new RerankerContext<>(searcher, null, query, null, queryString, queryTokens, null, searchArgs);
+      context = new RerankerContext<>(searcher, null, query, null, queryString, queryTokens, null, searchArgs);
     }
 
     ScoredDocuments hits = cascade.run(ScoredDocuments.fromTopDocs(rs, searcher), context);
