+++ src/test/java/io/anserini/index/BasicIndexOperationsTest.java
@@ -14,27 +14,14 @@
  * limitations under the License.
  */
 
-package io.anserini.integration;
+package io.anserini.index;
 
-import io.anserini.index.IndexReaderUtils;
-import org.apache.lucene.analysis.Analyzer;
-import org.apache.lucene.analysis.en.EnglishAnalyzer;
-import org.apache.lucene.codecs.FieldsProducer;
-import org.apache.lucene.document.Document;
-import org.apache.lucene.document.Field;
-import org.apache.lucene.document.FieldType;
-import org.apache.lucene.document.StringField;
-import org.apache.lucene.index.CodecReader;
+import io.anserini.IndexerTestBase;
 import org.apache.lucene.index.DirectoryReader;
-import org.apache.lucene.index.FilterCodecReader;
-import org.apache.lucene.index.IndexOptions;
 import org.apache.lucene.index.IndexReader;
-import org.apache.lucene.index.IndexWriter;
-import org.apache.lucene.index.IndexWriterConfig;
 import org.apache.lucene.index.LeafReader;
 import org.apache.lucene.index.MultiTerms;
 import org.apache.lucene.index.PostingsEnum;
-import org.apache.lucene.index.SlowCodecReaderWrapper;
 import org.apache.lucene.index.Term;
 import org.apache.lucene.index.Terms;
 import org.apache.lucene.index.TermsEnum;
@@ -48,89 +35,23 @@ import org.apache.lucene.search.TopDocs;
 import org.apache.lucene.search.similarities.BM25Similarity;
 import org.apache.lucene.store.Directory;
 import org.apache.lucene.store.FSDirectory;
-import org.apache.lucene.util.Accountable;
 import org.apache.lucene.util.BytesRef;
-import org.apache.lucene.util.LuceneTestCase;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 
 import java.io.IOException;
-import java.nio.file.Path;
-import java.util.Collection;
-import java.util.Iterator;
 
-public class IndexerTest extends LuceneTestCase {
-
-  private static Path tempDir1;
-  private static Path tempDir2;
-
-  // A very simple example of how to build an index.
-  private void buildTestIndex() throws IOException {
-    Directory dir = FSDirectory.open(tempDir1);
-
-    Analyzer analyzer = new EnglishAnalyzer();
-    IndexWriterConfig config = new IndexWriterConfig(analyzer);
-    config.setOpenMode(IndexWriterConfig.OpenMode.CREATE);
-
-    IndexWriter writer = new IndexWriter(dir, config);
-
-    FieldType textOptions = new FieldType();
-    textOptions.setIndexOptions(IndexOptions.DOCS_AND_FREQS_AND_POSITIONS);
-    textOptions.setStored(true);
-    textOptions.setTokenized(true);
-    textOptions.setStoreTermVectors(true);
-    textOptions.setStoreTermVectorPositions(true);
-
-    Document doc1 = new Document();
-    doc1.add(new StringField("id", "doc1", Field.Store.YES));
-    doc1.add(new Field("text", "here is some text here is some more text", textOptions));
-    writer.addDocument(doc1);
-
-    Document doc2 = new Document();
-    doc2.add(new StringField("id", "doc2", Field.Store.YES));
-    doc2.add(new Field("text", "more text", textOptions));
-    writer.addDocument(doc2);
-
-    Document doc3 = new Document();
-    doc3.add(new StringField("id", "doc3", Field.Store.YES));
-    doc3.add(new Field("text", "here is a test", textOptions));
-    writer.addDocument(doc3);
-
-    writer.commit();
-    writer.forceMerge(1);
-    writer.close();
-  }
-
-  @Before
-  @Override
-  public void setUp() throws Exception {
-    super.setUp();
-
-    tempDir1 = createTempDir();
-    tempDir2 = createTempDir();
-    buildTestIndex();
-  }
-
-  @After
-  @Override
-  public void tearDown() throws Exception {
-    // Call garbage collector for Windows compatibility
-    System.gc();
-    super.tearDown();
-  }
-
+public class BasicIndexOperationsTest extends IndexerTestBase {
 
   // A very simple example of how to iterate through terms in an index and dump out postings.
   private void dumpPostings(IndexReader reader) throws IOException {
     // This is how you iterate through terms in the postings list.
     LeafReader leafReader = reader.leaves().get(0).reader();
-    TermsEnum termsEnum = leafReader.terms("text").iterator();
+    TermsEnum termsEnum = leafReader.terms("contents").iterator();
     BytesRef bytesRef = termsEnum.next();
     while (bytesRef != null) {
       // This is the current term in the dictionary.
       String token = bytesRef.utf8ToString();
-      Term term = new Term("text", token);
+      Term term = new Term("contents", token);
 
       PostingsEnum postingsEnum;
 
@@ -144,7 +65,7 @@ public class IndexerTest extends LuceneTestCase {
       // How to dump out positional info as well, from test case at:
       // https://github.com/apache/lucene-solr/blob/master/lucene/core/src/test/org/apache/lucene/index/TestPostingsOffsets.java
       System.out.print(token + " (df = " + reader.docFreq(term) + "):");
-      postingsEnum = MultiTerms.getTermPostingsEnum(reader, "text", bytesRef);
+      postingsEnum = MultiTerms.getTermPostingsEnum(reader, "contents", bytesRef);
       while (postingsEnum.nextDoc() != DocIdSetIterator.NO_MORE_DOCS) {
         System.out.print(String.format(" (%s, %s)", postingsEnum.docID(), postingsEnum.freq()));
         System.out.print(" [");
@@ -169,15 +90,15 @@ public class IndexerTest extends LuceneTestCase {
     System.out.println("Dumping out postings...");
     dumpPostings(reader);
 
-    assertEquals(2, reader.docFreq(new Term("text", "here")));
-    assertEquals(2, reader.docFreq(new Term("text", "more")));
-    assertEquals(1, reader.docFreq(new Term("text", "some")));
-    assertEquals(1, reader.docFreq(new Term("text", "test")));
-    assertEquals(2, reader.docFreq(new Term("text", "text")));
+    assertEquals(2, reader.docFreq(new Term("contents", "here")));
+    assertEquals(2, reader.docFreq(new Term("contents", "more")));
+    assertEquals(1, reader.docFreq(new Term("contents", "some")));
+    assertEquals(1, reader.docFreq(new Term("contents", "test")));
+    assertEquals(2, reader.docFreq(new Term("contents", "text")));
 
     PostingsEnum postingsEnum;
     // here (df = 2): (0, 2) [0, 4] (2, 1) [0]
-    postingsEnum = MultiTerms.getTermPostingsEnum(reader, "text", new BytesRef("here"));
+    postingsEnum = MultiTerms.getTermPostingsEnum(reader, "contents", new BytesRef("here"));
     assertEquals(0, postingsEnum.nextDoc());
     assertEquals(0, postingsEnum.docID());
     assertEquals(2, postingsEnum.freq());
@@ -190,7 +111,7 @@ public class IndexerTest extends LuceneTestCase {
     assertEquals(DocIdSetIterator.NO_MORE_DOCS, postingsEnum.nextDoc());
 
     // more (df = 2): (0, 1) [7] (1, 1) [0]
-    postingsEnum = MultiTerms.getTermPostingsEnum(reader, "text", new BytesRef("more"));
+    postingsEnum = MultiTerms.getTermPostingsEnum(reader, "contents", new BytesRef("more"));
     assertEquals(0, postingsEnum.nextDoc());
     assertEquals(0, postingsEnum.docID());
     assertEquals(1, postingsEnum.freq());
@@ -202,7 +123,7 @@ public class IndexerTest extends LuceneTestCase {
     assertEquals(DocIdSetIterator.NO_MORE_DOCS, postingsEnum.nextDoc());
 
     // some (df = 1): (0, 2) [2, 6]
-    postingsEnum = MultiTerms.getTermPostingsEnum(reader, "text", new BytesRef("some"));
+    postingsEnum = MultiTerms.getTermPostingsEnum(reader, "contents", new BytesRef("some"));
     assertEquals(0, postingsEnum.nextDoc());
     assertEquals(0, postingsEnum.docID());
     assertEquals(2, postingsEnum.freq());
@@ -211,7 +132,7 @@ public class IndexerTest extends LuceneTestCase {
     assertEquals(DocIdSetIterator.NO_MORE_DOCS, postingsEnum.nextDoc());
 
     // test (df = 1): (2, 1) [3]
-    postingsEnum = MultiTerms.getTermPostingsEnum(reader, "text", new BytesRef("test"));
+    postingsEnum = MultiTerms.getTermPostingsEnum(reader, "contents", new BytesRef("test"));
     assertEquals(2, postingsEnum.nextDoc());
     assertEquals(2, postingsEnum.docID());
     assertEquals(1, postingsEnum.freq());
@@ -219,7 +140,7 @@ public class IndexerTest extends LuceneTestCase {
     assertEquals(DocIdSetIterator.NO_MORE_DOCS, postingsEnum.nextDoc());
 
     // text (df = 2): (0, 2) [3, 8] (1, 1) [1]
-    postingsEnum = MultiTerms.getTermPostingsEnum(reader, "text", new BytesRef("text"));
+    postingsEnum = MultiTerms.getTermPostingsEnum(reader, "contents", new BytesRef("text"));
     assertEquals(0, postingsEnum.nextDoc());
     assertEquals(0, postingsEnum.docID());
     assertEquals(2, postingsEnum.freq());
@@ -245,13 +166,13 @@ public class IndexerTest extends LuceneTestCase {
     // Iterate through the document vectors
     for (int i=0; i<numDocs; i++) {
       System.out.println(reader.document(i));
-      Terms terms = reader.getTermVector(i, "text");
+      Terms terms = reader.getTermVector(i, "contents");
       TermsEnum te = terms.iterator();
 
       // For this document, iterate through the terms.
       Term term;
       while (te.next() != null) {
-        term = new Term("text", te.term());
+        term = new Term("contents", te.term());
         long tf = te.totalTermFreq();
         // Print out the term and its term frequency
         System.out.println(term.bytes().utf8ToString() + " " + tf);
@@ -274,18 +195,18 @@ public class IndexerTest extends LuceneTestCase {
       String docid = reader.document(i).getField("id").stringValue();
       System.out.println(reader.document(i));
       System.out.println(i+ ": " + docid);
-      Terms terms = reader.getTermVector(i, "text");
+      Terms terms = reader.getTermVector(i, "contents");
       TermsEnum te = terms.iterator();
 
       // For this document, iterate through the terms.
       while (te.next() != null) {
-        String term = new Term("text", te.term()).bytes().utf8ToString();
+        String term = new Term("contents", te.term()).bytes().utf8ToString();
         long tf = te.totalTermFreq();
 
         // The way to compute the BM25 score is to issue a query with the exact docid and the
         // term in question, and look at the retrieval score.
-        Query filterQuery = new TermQuery(new Term("docid", docid)); // the docid
-        Query termQuery =  new TermQuery(new Term("text", term));    // the term
+        Query filterQuery = new TermQuery(new Term("id", docid)); // the docid
+        Query termQuery =  new TermQuery(new Term("contents", term));    // the term
         BooleanQuery.Builder builder = new BooleanQuery.Builder();   // must have both
         builder.add(filterQuery, BooleanClause.Occur.MUST);
         builder.add(termQuery, BooleanClause.Occur.MUST);
@@ -295,129 +216,6 @@ public class IndexerTest extends LuceneTestCase {
         // The BM25 weight is the maxScore
         System.out.println(term + " " + tf + " " + (rs.scoreDocs.length == 0 ? Float.NaN : rs.scoreDocs[0].score));
       }
-    }
-  }
-
-  @Test
-  public void testDocidConversion() throws Exception {
-    Directory dir = FSDirectory.open(tempDir1);
-    IndexReader reader = DirectoryReader.open(dir);
-
-    assertEquals("doc1", IndexReaderUtils.convertLuceneDocidToDocid(reader, 0));
-    assertEquals("doc2", IndexReaderUtils.convertLuceneDocidToDocid(reader, 1));
-    assertEquals("doc3", IndexReaderUtils.convertLuceneDocidToDocid(reader, 2));
-
-    assertEquals(0, IndexReaderUtils.convertDocidToLuceneDocid(reader, "doc1"));
-    assertEquals(1, IndexReaderUtils.convertDocidToLuceneDocid(reader, "doc2"));
-    assertEquals(2, IndexReaderUtils.convertDocidToLuceneDocid(reader, "doc3"));
-  }
-
-  @Test
-  public void testCloneIndex() throws Exception {
-    System.out.println("Cloning index:");
-    Directory dir1 = FSDirectory.open(tempDir1);
-    IndexReader reader = DirectoryReader.open(dir1);
-
-    Directory dir2 = FSDirectory.open(tempDir2);
-    IndexWriterConfig config = new IndexWriterConfig(new EnglishAnalyzer());
-    config.setOpenMode(IndexWriterConfig.OpenMode.CREATE);
-    IndexWriter writer = new IndexWriter(dir2, config);
-
-    LeafReader leafReader = reader.leaves().get(0).reader();
-    CodecReader codecReader = SlowCodecReaderWrapper.wrap(leafReader);
-    writer.addIndexes(new MyFilterCodecReader(codecReader));
-    writer.commit();
-    writer.forceMerge(1);
-    writer.close();
-
-    reader.close();
-
-    // Open up the cloned index and verify it.
-    reader = DirectoryReader.open(dir2);
-    assertEquals(3, reader.numDocs());
-    assertEquals(1, reader.leaves().size());
-
-    System.out.println("Dumping out postings...");
-    dumpPostings(reader);
-
-    assertEquals(2, reader.docFreq(new Term("text", "here")));
-    assertEquals(2, reader.docFreq(new Term("text", "more")));
-    assertEquals(1, reader.docFreq(new Term("text", "some")));
-    assertEquals(1, reader.docFreq(new Term("text", "test")));
-    assertEquals(2, reader.docFreq(new Term("text", "text")));
-
-    reader.close();
-  }
-
-  // Custom class so we can intercept calls and potentially alter behavior.
-  private static class MyFilterCodecReader extends FilterCodecReader {
-    final private CodecReader in;
-
-    public MyFilterCodecReader(CodecReader in) {
-      super(in);
-      this.in = in;
-    }
-
-    @Override
-    public FieldsProducer getPostingsReader() {
-      System.out.println("Getting custom postings reader...");
-      return new MyFieldsProducer(in.getPostingsReader());
-    }
-
-    @Override
-    public IndexReader.CacheHelper getCoreCacheHelper() {
-      throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public IndexReader.CacheHelper getReaderCacheHelper() {
-      throw new UnsupportedOperationException();
-    }
-  }
-
-  // Custom class so we can intercept calls and potentially alter behavior.
-  private static class MyFieldsProducer extends FieldsProducer {
-    final private FieldsProducer fieldsProducer;
-
-    public MyFieldsProducer(FieldsProducer fieldsProducer) {
-      this.fieldsProducer = fieldsProducer;
-    }
-
-    @Override
-    public void close() throws IOException {
-      fieldsProducer.close();
-    }
-
-    @Override
-    public void checkIntegrity() throws IOException {
-      fieldsProducer.iterator();
-    }
-
-    @Override
-    public Iterator<String> iterator() {
-      return fieldsProducer.iterator();
-    }
-
-    @Override
-    public Terms terms(String s) throws IOException {
-      System.out.println("Intercepting call to method 'terms': " + s);
-
-      return fieldsProducer.terms(s);
-    }
-
-    @Override
-    public int size() {
-      return fieldsProducer.size();
-    }
-
-    @Override
-    public long ramBytesUsed() {
-      return fieldsProducer.ramBytesUsed();
-    }
-
-    @Override
-    public Collection<Accountable> getChildResources() {
-      return fieldsProducer.getChildResources();
     }
   }
 }
